+++ src/main/java/net/silentchaos512/gear/item/Flaxseeds.java
@@ -4,7 +4,6 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemSeeds;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
@@ -12,13 +11,11 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.EnumPlantType;
-import net.minecraftforge.common.MinecraftForge;
 import net.silentchaos512.gear.init.ModBlocks;
 
 public class Flaxseeds extends ItemSeeds {
     public Flaxseeds() {
         super(ModBlocks.flaxPlant, Blocks.FARMLAND);
-        MinecraftForge.addGrassSeed(new ItemStack(this), 5);
     }
 
     @Override