+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Config to set flaxseed drop weight, set 0 to disable [#14]
 ### Changed
 - Shovels no longer make paths while sneaking (more consistent block placing)
 - Convert most remaining recipe types to JSON. Should not have any effect, just 1.13 prep.